Defender refactoring
Move code related to lock and unlock user,ip address.
Update usages in class_views.py
