+++ core/class_views.py
@@ -12,8 +12,7 @@ from django.shortcuts import redirect
 from django.views.generic.list import ListView
 from django.utils.timezone import now
 from django.contrib.auth.forms import AuthenticationForm
-from pleio_account import utils
-from defender import config
+from accountlockout import users, utils
 
 class PleioLoginView(LoginView):
 
@@ -28,16 +27,14 @@ class PleioLoginView(LoginView):
     )
 
     def get_context_data(self, **kwargs):
-        attempts = utils.get_user_attempts(self.request)
-        get_username = utils.get_username_from_request(self.request)
-        time = int(config.COOLOFF_TIME / 60)
-        attempts_left = (config.FAILURE_LIMIT - utils.get_user_attempts(self.request))
-        kwargs = dict(kwargs,attempts=attempts,username=get_username, time=time, attempts_left=attempts_left)
+        attempts = users.get_user_attempts((self.request);
+        get_username = users.get_username_from_request(self.request);
+        kwargs = dict(kwargs,attempts=attempts,username=get_username, time=utils.get_time(), attempts_left=utils.get_attemps_left(self.request));
 
-        context = super(PleioLoginView, self).get_context_data(**kwargs)
-        next = self.request.GET.get('next')
+        context = super(PleioLoginView, self).get_context_data(**kwargs);
+        next = self.request.GET.get('next');
         if next:
-            context['next'] = next
+            context['next'] = next;
 
         self.set_partner_site_info()
